+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.4] - 2021-05-18
 ### Added
 - Partial implementation of the Primaries swatch - [#24](https://github.com/dartoos-dev/eo_color/issues/24).
 - Bluish color palettes and swatches - [#29](https://github.com/dartoos-dev/eo_color/issues/29).